refactor: refactor auth pages
