+++ frontend/src/pages/auth/Register.tsx
@@ -1,109 +1,17 @@
-import { useState } from "react";
-import { Link } from "react-router-dom";
-import {
-  Box,
-  Button,
-  Container,
-  TextField,
-  Typography,
-  Paper,
-  CircularProgress,
-} from "@mui/material";
+import AuthForm from "../../components/AuthForm";
 import { useAuth } from "../../contexts/AuthContext";
 
 export default function Register() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
   const { register, loading } = useAuth();
 
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    await register(email, password);
-  };
-
   return (
-    <Container component="main" maxWidth="xs">
-      <Box
-        sx={{
-          marginTop: 8,
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-        }}
-      >
-        <Paper
-          elevation={3}
-          sx={{
-            padding: 4,
-            display: "flex",
-            flexDirection: "column",
-            alignItems: "center",
-            width: "100%",
-          }}
-        >
-          <Typography
-            component="h1"
-            variant="h5"
-            fontWeight={700}
-            color="primary"
-          >
-            DynaPredict
-          </Typography>
-          <Typography component="h2" variant="h6" mt={2} mb={3}>
-            Crie uma conta
-          </Typography>
-          <Box
-            component="form"
-            onSubmit={handleSubmit}
-            sx={{ mt: 1, width: "100%" }}
-          >
-            <TextField
-              margin="normal"
-              required
-              fullWidth
-              id="email"
-              label="Email Address"
-              name="email"
-              autoComplete="email"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              autoFocus
-            />
-            <TextField
-              margin="normal"
-              required
-              fullWidth
-              name="password"
-              label="Password"
-              type="password"
-              id="password"
-              autoComplete="new-password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-            />
-            <Button
-              type="submit"
-              fullWidth
-              variant="contained"
-              sx={{ mt: 3, mb: 2, py: 1.5 }}
-              disabled={loading}
-            >
-              {loading ? <CircularProgress size={24} /> : "Register"}
-            </Button>
-            <Box sx={{ textAlign: "center", mt: 2 }}>
-              <Typography
-                variant="body2"
-                sx={{ display: "flex", gap: "4px", justifyContent: "center" }}
-              >
-                Já possui uma conta?
-                <Link to="/auth/register" style={{ color: "primary" }}>
-                  Entrar
-                </Link>
-              </Typography>
-            </Box>
-          </Box>
-        </Paper>
-      </Box>
-    </Container>
+    <AuthForm
+      title="Crie uma conta"
+      buttonText="Registrar"
+      linkText="Já possui uma conta?"
+      linkTo="/auth/login"
+      onSubmit={register}
+      loading={loading}
+    />
   );
 }
